pytest-bdd-sample/harness/pyproject.toml
[tool.poetry]
name = "pytest-bdd-sample"
version = "0.1.0"
description = ""
authors = ["Robert Bradley <rbradaws@amazon.co.uk>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^7.4.3"
pytest-bdd = "^7.0.0"
selenium = "^4.15.0"
pyhamcrest = "^2.1.0"
pytest-xdist = "^3.3.1"
coverage = "^7.3.2"
black = "^23.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
